Call SQLSetConnectAttr properly
Evidently we must pass the value of the connectTimeout or loginTimeout to
SQLSetConnectAttr, not a pointer to the value. This is super counter-intuitive
because we have to cast the value to a SQLPOINTER in order to get it to even
compile. In order to avoid compiler warnings we must cast to size_t before
casting to SQLPOINTER.

This will properly timeout connections when attempting to connect to a
non-running server; at least with the FreeTDS driver.