Fix test-domains-open failure.
Store the active domain on the database and connection objects and
use the `Nan::Callback::Call()` overload that enters the domain
before making the callback.

This does the bare minimum to make the test pass.  Better domains
support would need a complete audit of the code base, but as domains
are a deprecated feature anyway, I don't think it's worth the effort.