Fix endian bug in stored procedure output parms
When calling stored procedures with output parameters, the output is
always blank. This is due to using the wrong size field when querying
SQLProcedureColumns to retrieve the column size for each parameter. This
field is defined by ODBC as an INTEGER, but the pointer is read from a
SMALLINT instead. This results in the size being read from only the top
2 bytes of the 4 byte value. On little endian systems, this will be OK
so long as the size is < 32k. On big endian systems it will always be 0
unless the size is > 32k or negative; regardless it will always be
*wrong*.

Signed-off-by: Kevin Adler <kadler@us.ibm.com>

Fixes #103