Fix callProcedure core dump with large parameters (#318)
When allocating the buffer for the procedure parameters, the size is
cast to an SQLSMALLINT, ie. signed short. For CHAR parameters longer
than 8191, the SQLSMALLINT overflows and become negative after we
multiply by 4 to account for possible UTF-8 expansion. The negative
value is then passed to new[]() which throws a std::bad_alloc.

I'm not sure why an SQLSMALLINT was used here, since it is only used
temporarily before being stored in BufferLength, which is an SQLLEN so
use that type instead. This temp variable is only needed because we need
the original BufferLength to memcpy during reallocation of
ParameterValuePtr.

In addition, after reallocation of ParameterValuePtr, the original
pointer is never freed, leading to a memory leak. A lot of this code
would be cleaned up if we replaced raw pointers with C++ containers like
std::vector (would just be a single call to resize()), but that's a much
larger rewrite.

Fixes #313

Signed-off-by: Kevin Adler <kadler@us.ibm.com>